+++ debugger/emmy/emmyHelper.lua
@@ -187,7 +187,14 @@ elseif xlua then
     emmy = xluaDebugger
 end
 
-_G.emmyHelper = emmy
-if _G.emmyHelperInit then
-    _G.emmyHelperInit()
+local emmyHelper = rawget(_G, "emmyHelper")
+if emmyHelper == nil then
+    rawset(_G, 'emmyHelper', emmy)
+else
+    emmyHelper.queryVariable = emmy.queryVariable
+end
+
+local emmyHelperInit = rawget(_G, 'emmyHelperInit')
+if emmyHelperInit then
+    emmyHelperInit()
 end