+++ src/annotator.ts
@@ -4,26 +4,46 @@ import { AnnotatorType } from './notifications';
 import { LanguageClient } from 'vscode-languageclient';
 import * as notifications from "./notifications";
 
-let D_PARAM:vscode.TextEditorDecorationType;
-let D_GLOBAL:vscode.TextEditorDecorationType;
-let D_DOC_TYPE:vscode.TextEditorDecorationType;
-let D_UPVALUE:vscode.TextEditorDecorationType;
 
-function createDecoration(key: string, config: vscode.DecorationRenderOptions|undefined = undefined): vscode.TextEditorDecorationType {
+let D_PARAM: vscode.TextEditorDecorationType;
+let D_GLOBAL: vscode.TextEditorDecorationType;
+let D_DOC_TYPE: vscode.TextEditorDecorationType;
+let D_UPVALUE: vscode.TextEditorDecorationType;
+let D_NOTUSE: vscode.TextEditorDecorationType;
+let D_PARAMHINT: vscode.TextEditorDecorationType;
+let D_LOCALHINT: vscode.TextEditorDecorationType;
+
+function createDecoration(key: string, config: vscode.DecorationRenderOptions | undefined = undefined): vscode.TextEditorDecorationType {
     let color = vscode.workspace.getConfiguration("emmylua").get(key);
     config = config || {};
-    if (typeof(color) === 'string') {
-        config.light = { color: color};
-        config.dark = { color: color};
+    if (typeof (color) === 'string') {
+        config.light = { color: color };
+        config.dark = { color: color };
     }
     return vscode.window.createTextEditorDecorationType(config);
 }
 
 function updateDecorations() {
+    // 各种方式更新时之前的decoration没有dispose导致重复渲染
+    if (D_PARAM) {
+        D_PARAM.dispose();
+        D_GLOBAL.dispose();
+        D_DOC_TYPE.dispose();
+        D_UPVALUE.dispose();
+        D_NOTUSE.dispose();
+        D_PARAMHINT.dispose();
+        D_LOCALHINT.dispose();
+    }
+
     D_PARAM = createDecoration("colors.parameter");
     D_GLOBAL = createDecoration("colors.global");
     D_DOC_TYPE = createDecoration("colors.doc_type");
-    D_UPVALUE = createDecoration("", { textDecoration: "underline" });
+    D_UPVALUE = createDecoration("colors.", { 
+        textDecoration: "underline;text-decoration-color:#a8c023;"
+    });
+    D_NOTUSE = createDecoration("colors.not_use", {});
+    D_PARAMHINT = vscode.window.createTextEditorDecorationType({});
+    D_LOCALHINT = vscode.window.createTextEditorDecorationType({});
 }
 
 export function onDidChangeConfiguration(client: LanguageClient) {
@@ -44,20 +64,30 @@ function requestAnnotatorsImpl(editor: vscode.TextEditor, client: LanguageClient
         updateDecorations();
     }
 
-    let params:notifications.AnnotatorParams = { uri: editor.document.uri.toString() };
+    let params: notifications.AnnotatorParams = { uri: editor.document.uri.toString() };
     client.sendRequest<notifications.IAnnotator[]>("emmy/annotator", params).then(list => {
-        let map: Map<AnnotatorType, vscode.Range[]> = new Map();
+        let map: Map<AnnotatorType, notifications.RenderRange[]> = new Map();
         map.set(AnnotatorType.DocType, []);
         map.set(AnnotatorType.Param, []);
         map.set(AnnotatorType.Global, []);
         map.set(AnnotatorType.Upvalue, []);
+        map.set(AnnotatorType.NotUse, []);
+        map.set(AnnotatorType.ParamHint, []);
+        map.set(AnnotatorType.LocalHint, []);
 
         list.forEach(data => {
             let uri = vscode.Uri.parse(data.uri);
+            let uriSet = new Set<string>();
+            // 而vscode 在diff，分屏以及其他一些情况下可以获得多个相同的uri
             vscode.window.visibleTextEditors.forEach((editor) => {
                 let docUri = editor.document.uri;
+                if (uriSet.has(docUri.path)) {
+                    return;
+                }
+                uriSet.add(docUri.path)
+
                 if (uri.path.toLowerCase() === docUri.path.toLowerCase()) {
-                    var list = map.get(data.type);
+                    let list = map.get(data.type);
                     if (list === undefined) {
                         list = data.ranges;
                     } else {
@@ -73,19 +103,91 @@ function requestAnnotatorsImpl(editor: vscode.TextEditor, client: LanguageClient
     });
 }
 
-function updateAnnotators(editor: vscode.TextEditor, type: AnnotatorType, ranges: vscode.Range[]) {
+function updateAnnotators(editor: vscode.TextEditor, type: AnnotatorType, renderRanges: notifications.RenderRange[]) {
     switch (type) {
         case AnnotatorType.Param:
-        editor.setDecorations(D_PARAM, ranges);
-        break;
+            editor.setDecorations(D_PARAM, renderRanges.map(e => e.range));
+            break;
         case AnnotatorType.Global:
-        editor.setDecorations(D_GLOBAL, ranges);
-        break;
+            editor.setDecorations(D_GLOBAL, renderRanges.map(e => e.range));
+            break;
         case AnnotatorType.DocType:
-        editor.setDecorations(D_DOC_TYPE, ranges);
-        break;
+            editor.setDecorations(D_DOC_TYPE, renderRanges.map(e => e.range));
+            break;
         case AnnotatorType.Upvalue:
-        editor.setDecorations(D_UPVALUE, ranges);
-        break;
+            editor.setDecorations(D_UPVALUE, renderRanges.map(e => e.range));
+            break;
+        case AnnotatorType.NotUse:
+            editor.setDecorations(D_NOTUSE, renderRanges.map(e => e.range));
+            break;
+        case AnnotatorType.ParamHint: {
+            let vscodeRenderRanges: vscode.DecorationOptions[] = []
+            renderRanges.forEach(renderRange => {
+                if (renderRange.hint && renderRange.hint !== "") {
+                    vscodeRenderRanges.push({
+                        range: renderRange.range,
+                        renderOptions: {
+                            light: {
+                                before: {
+                                    contentText: ` ${renderRange.hint}: `,
+                                    color: "#888888",
+                                    backgroundColor: '#EEEEEE;border-radius: 2px;',
+                                    fontWeight: '400; font-size: 12px; line-height: 1;',
+                                    margin: "1px",
+                                }
+                            },
+                            dark: {
+                                before: {
+                                    contentText: ` ${renderRange.hint}: `,
+                                    color: "#888888",
+                                    backgroundColor: '#333333;border-radius: 2px;',
+                                    fontWeight: '400; font-size: 12px; line-height: 1;',
+                                    margin: "1px",  
+                                    
+                                }
+                            }
+                        }
+                    });
+
+                }
+            });
+
+            editor.setDecorations(D_PARAMHINT, vscodeRenderRanges);
+            break;
+        }
+        case AnnotatorType.LocalHint: {
+            let vscodeRenderRanges: vscode.DecorationOptions[] = []
+            renderRanges.forEach(renderRange => {
+                if (renderRange.hint && renderRange.hint !== "") {
+                    vscodeRenderRanges.push({
+                        range: renderRange.range,
+                        renderOptions: {
+                            light: {
+                                after: {
+                                    contentText: `:${renderRange.hint}`,
+                                    color: "#888888",
+                                    backgroundColor: '#EEEEEE;border-radius: 2px;',
+                                    fontWeight: '400; font-size: 12px; line-height: 1;',
+                                    margin: "3px",
+                                }
+                            },
+                            dark: {
+                                after: {
+                                    contentText: `:${renderRange.hint}`,
+                                    color: "#888888",
+                                    backgroundColor: '#333333;border-radius: 2px;',
+                                    fontWeight: '400; font-size: 12px; line-height: 1;',
+                                    margin: "3px",
+                                }
+                            }
+                        }
+                    });
+
+                }
+            });
+
+            editor.setDecorations(D_LOCALHINT, vscodeRenderRanges);
+            break;
+        }
     }
 }