+++ src/debugger/EmmyAttachDebugSession.ts
@@ -5,11 +5,13 @@ import { EmmyDebugSession } from "./EmmyDebugSession";
 import { OutputEvent } from "vscode-debugadapter";
 import { DebugProtocol } from "vscode-debugprotocol";
 
+
 interface EmmyAttachDebugArguments extends DebugProtocol.AttachRequestArguments {
-	extensionPath: string;
+    extensionPath: string;
     sourcePaths: string[];
     ext: string[];
     pid: number;
+    captureLog?: boolean;
 }
 
 enum WinArch {
@@ -19,6 +21,7 @@ enum WinArch {
 export class EmmyAttachDebugSession extends EmmyDebugSession {
 
     private pid = 0;
+    private captureLog?: boolean;
 
     private getPort(pid: number): number {
         var port = pid;
@@ -31,20 +34,23 @@ export class EmmyAttachDebugSession extends EmmyDebugSession {
         this.extensionPath = args.extensionPath;
         this.ext = args.ext;
         this.pid = args.pid;
+        this.captureLog = args.captureLog;
+
         await this.attach();
+        
         // send resp
         const client = net.connect(this.getPort(args.pid), 'localhost')
-        .on('connect', () => {
-            this.sendResponse(response);
-            this.onConnect(client);
-            this.readClient(client);
-            this.sendMessage({ cmd: proto.MessageCMD.StartHookReq });
-        })
-        .on('error', err => {
-            response.success = false;
-            response.message = `${err}`;
-            this.sendResponse(response);
-        });
+            .on('connect', () => {
+                this.sendResponse(response);
+                this.onConnect(client);
+                this.readClient(client);
+                this.sendMessage({ cmd: proto.MessageCMD.StartHookReq });
+            })
+            .on('error', err => {
+                response.success = false;
+                response.message = `${err}`;
+                this.sendResponse(response);
+            });
         this.client = client;
     }
 
@@ -55,16 +61,17 @@ export class EmmyAttachDebugSession extends EmmyDebugSession {
             'arch_pid',
             `${this.pid}`
         ];
+
         return new Promise<WinArch>((r, c) => {
             cp.exec(args.join(" "), { cwd: cwd })
-            .on('close', (code) => {
-                r(code === 0 ? WinArch.X64 : WinArch.X86);
-            })
-            .on('error', c);
+                .on('close', (code) => {
+                    r(code === 0 ? WinArch.X64 : WinArch.X86);
+                })
+                .on('error', c);
         });
     }
 
-    private async attach() {
+    private async attach(): Promise<void> {
         const arch = await this.detectArch();
         const archName = arch === WinArch.X64 ? 'x64' : 'x86';
         const cwd = `${this.extensionPath}/debugger/emmy/windows/${archName}`;
@@ -78,18 +85,33 @@ export class EmmyAttachDebugSession extends EmmyDebugSession {
             '-dll',
             'emmy_hook.dll'
         ];
+        if(this.captureLog){
+            args.push("-capture-log");
+        }
+
         return new Promise<void>((r, c) => {
             cp.exec(args.join(" "), { cwd: cwd }, (err, stdout, stderr) => {
                 this.sendEvent(new OutputEvent(stdout));
             })
-            .on('close', (code) => {
-                if (code === 0) {
-                    r();
-                }
-                else {
-                    c(`Exit code = ${code}`);
-                }
-            });
+                .on('close', (code) => {
+                    if (code === 0) {
+                        if(this.captureLog){
+                            const captureArgs = [
+                                "emmy_tool.exe",
+                                "receive_log",
+                                "-p",
+                                `${this.pid}`,
+                            ] 
+                            cp.spawn(`wt`, captureArgs, {
+                                cwd: cwd
+                            });
+                        }
+                        r();
+                    }
+                    else {
+                        c(`Exit code = ${code}`);
+                    }
+                });
         });
     }