+++ src/notifications.ts
@@ -9,11 +9,20 @@ export enum AnnotatorType {
     Global,
     DocType,
     Upvalue,
+    NotUse,
+    ParamHint,
+    LocalHint,
 }
+
+export interface RenderRange{
+    range: vscode.Range;
+    hint: string;
+}
+
 
 export interface IAnnotator {
     uri: string;
-    ranges: vscode.Range[];
+    ranges: RenderRange[];
     type: AnnotatorType;
 }