+++ CHANGELOG.md
@@ -1,7 +1,50 @@
 # Change Log
-All notable changes to the "emmylua" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 0.4.0
 
-## [Unreleased]
-- Initial release
+恢复更新
+
+代码提示：
+* 支持lua5.4语法
+* 支持通过标注参数类型为函数，补全匿名回调函数
+* 修改函数签名显示
+* 支持参数类型诊断(实验性)
+* 支持枚举类型标注enum(实验性)
+
+代码渲染：
+* 支持param hint
+* 支持local hint
+* 支持vararg hint
+* 支持未使用变量渲染为灰色
+* upvalue修改为黄色下划线
+
+
+代码折叠：
+* 支持通过--region和--endregion实现区块折叠
+* 支持连续require语句折叠，连续require语句被识别为import区块，可以通过相关vscode设置直接折叠打开文件的require语句
+
+lua语言:
+* 移除lua语言定义配置文件直接采用vscode新的内建的lua语言定义
+
+格式化:
+* 移除现行的格式化算法，EmmyLua插件不再提供格式化算法
+* 格式化交由插件EmmyLuaCodeStyle实现
+
+调试:
+* 支持条件断点，命中次数断点，日志记录点
+* 支持lua5.4调试
+* 支持苹果m1下的调试
+* 附加调试支持指定匹配进程名
+* 调试提供debug inline variable 
+* 在windows上提供launch调试，launch调试支持使用windows terminal
+
+JVM:
+* kotlin修改为1.6版本，gradle修改为6.1.1版本
+* jdk推荐使用jdk11
+* JVM参数移除CMS采用G1GC
+
+BUG:
+* 修复键入代码的过程中染色错误的问题
+* 修复因为外部更新没有即时刷新索引的问题
+* 修复函数签名错误
+* 重写调试器修复常见崩溃问题